重构：抽取工具类
