+++ src/core/interfaces/IRelation.ts
@@ -1,0 +1,4 @@
+interface IRelation {
+    property: string;
+    elements: JQuery<HTMLElement>[];
+}
