+++ src/core/vm.ts
@@ -1,23 +1,26 @@
 /// <reference path="statics.ts" />
 /// <reference path="component.ts" />
+/// <reference path="util.ts" />
+/// <reference path="relationCollection.ts" />
 
 namespace Pandyle {
     export class VM<T> {
         protected _data: T;
-        private _relations: relation[];
         private _root: JQuery<HTMLElement>;
         private _methods: object;
         private _filters: object;
-        private _converters: object;
+        public _converters: object;
         private _variables: object;
         private _defaultAlias: object;
+
+        private _relationCollection: IRelationCollection;
+        private _util: Util<T>;
 
         private static _uid = 1;
 
         constructor(element: JQuery<HTMLElement>, data: T, autoRun: boolean = true) {
             this._data = $.extend({}, data);
             this._root = element;
-            this._relations = [];
             this._methods = {};
             this._filters = {};
             this._converters = {};
@@ -32,6 +35,10 @@ namespace Pandyle {
                     property: '@window'
                 }
             }
+            
+            this._util = Util.CreateUtil(this);
+            this._relationCollection = relationCollection.CreateRelationCollection(this._util);
+
             if (autoRun) {
                 this.run();
             }
@@ -57,13 +64,9 @@ namespace Pandyle {
                 }
                 target[lastProperty] = _newData[key];
                 if ($.isArray(target[lastProperty])) {
-                    for (let i = this._relations.length - 1; i >= 0; i--) {
-                        if (this.isChild(key, this._relations[i].property)) {
-                            this._relations.splice(i, 1);
-                        }
-                    }
+                    this._relationCollection.removeChildren(key);
                 }
-                let relation = this._relations.filter(value => this.isSelfOrChild(key, value.property));
+                let relation = this._relationCollection.findSelfOrChild(key);
                 if (relation.length > 0) {
                     relation[0].elements.forEach(ele => {
                         this.render(ele);
@@ -80,11 +83,11 @@ namespace Pandyle {
                 case 'array':
                     return param.map(value => this.get(value));
                 case 'string':
-                    return this.getValue(this._root, param, this._data);
+                    return this._util.getValue(this._root, param, this._data);
                 case 'object':
                     let result = {};
                     for (let key in param) {
-                        result[key] = this.getValue(this._root, param[key], this._data);
+                        result[key] = this._util.getValue(this._root, param[key], this._data);
                     }
                     return result;
                 default:
@@ -128,7 +131,7 @@ namespace Pandyle {
                 element.data('alias', alias);
             }
             data = element.data('context');
-            this.setAlias(element, parentProperty, data);
+            this._util.setAlias(element, parentProperty, data);
             if(!this.bindIf(ele, parentProperty)){
                 return;
             }
@@ -225,19 +228,19 @@ namespace Pandyle {
             if (element.attr('p-context')) {
                 let data = element.data('context');
                 let expression = element.attr('p-context');
-                let divided = this.dividePipe(expression);
+                let divided = this._util.dividePipe(expression);
                 let property = divided.property;
                 let method = divided.method;
-                let target: any = this.calcu(property, element, data);
+                let target: any = this._util.calcu(property, element, data);
                 if (method) {
-                    target = this.convert(method, $.extend({}, target));
+                    target = this._util.convert(method, $.extend({}, target));
                 }
                 let fullProp = property;
                 if (parentProperty !== '') {
                     fullProp = parentProperty + '.' + property;
                 }
-                this.setAlias(element, fullProp, target);
-                this.setRelation(property, $(ele), parentProperty);
+                this._util.setAlias(element, fullProp, target);
+                this._relationCollection.setRelation(property, $(ele), parentProperty);
                 this.renderChild(ele, target, fullProp);
             }
         }
@@ -271,16 +274,16 @@ namespace Pandyle {
             let $this = this;
             if (element.attr('p-each')) {
                 let expression = element.attr('p-each').replace(/\s/g, '');
-                let divided = this.dividePipe(expression);
+                let divided = this._util.dividePipe(expression);
                 let property = divided.property;
                 let method = divided.method;
-                let target: any[] = this.calcu(property, element, data);
+                let target: any[] = this._util.calcu(property, element, data);
                 if (method) {
                     target = this.filter(method, target);
                 }
                 if (!element.data('pattern')) {
                     element.data('pattern', element.html());
-                    this.setRelation(property, element, parentProperty);
+                    this._relationCollection.setRelation(property, element, parentProperty);
                 };
                 let fullProp = property;
                 if (parentProperty !== '') {
@@ -306,10 +309,10 @@ namespace Pandyle {
                     element.data('uid', VM._uid++);
                 }
                 let expression = element.attr('p-for').replace(/\s/g, '');
-                let divided = this.dividePipe(expression);
+                let divided = this._util.dividePipe(expression);
                 let property = divided.property;
                 let method = divided.method;
-                let target: any[] = this.calcu(property, element, data);
+                let target: any[] = this._util.calcu(property, element, data);
                 if (method) {
                     target = this.filter(method, target);
                 }
@@ -339,7 +342,7 @@ namespace Pandyle {
                         newSibling.data('pattern', htmlText);
                         newSibling.attr('p-for', expression);
                         newSibling.data('context', data);
-                        $this.setRelation(property, newSibling, parentProperty);
+                        $this._relationCollection.setRelation(property, newSibling, parentProperty);
                     } else {
                         newSibling.attr('uid', element.data('uid'));
                     }
@@ -388,223 +391,30 @@ namespace Pandyle {
             }
             let result = pattern.replace(reg, ($0, $1) => {
                 if (!related) {
-                    this.setRelation($1, element, parentProperty);
+                    this._relationCollection.setRelation($1, element, parentProperty);
                     element.data('binding')[prop].related = true;
                 }
-                return this.getValue(element, $1, data);
+                return this._util.getValue(element, $1, data);
             });
             return result;
         }
 
-        /**
-         * 
-         * @param property 数据字段
-         * @param element 跟该字段绑定的对象
-         * @param parentProperty 父级字段的名称
-         */
-        private setRelation(property: string, element: JQuery<HTMLElement>, parentProperty) {
-            if (/^@.*/.test(property)) {
-                property = property.replace(/@(\w+)?/, ($0, $1) => {
-                    return this.getAliasProperty(element, $1);
-                })
-            } else if (parentProperty != '') {
-                property = parentProperty + '.' + property;
-            }
-            if (/^\./.test(property)) {
-                property = property.substr(1);
-            }
-            let relation = this._relations.filter(value => value.property === property);
-            if (relation.length == 0) {
-                this._relations.push({
-                    property: property,
-                    elements: [element]
-                });
-            } else {
-                if (relation[0].elements.indexOf(element) < 0) {
-                    relation[0].elements.push(element);
-                }
-            }
-        }
-
-        /**
-         * 从字段名称判断一个字段是否跟目标字段一致或属于目标字段的子字段
-         * @param property 目标字段
-         * @param subProperty 子字段
-         */
-        private isSelfOrChild(property: string, subProperty: string) {
-            property = property.replace(/[\[\]\(\)\.]/g, $0 => {
-                return '\\' + $0;
-            })
-            let reg = new RegExp('^' + property + '$' + '|' + '^' + property + '[\\[\\.]\\w+');
-            return reg.test(subProperty);
-        }
-
-        /**
-         * 从字段名称判断一个字段是否属于目标字段的子字段
-         * @param property 目标字段
-         * @param subProperty 子字段
-         */
-        private isChild(property: string, subProperty: string) {
-            let reg = new RegExp('^' + property + '[\\[\\.]\\w+');
-            return reg.test(subProperty);
-        }
-
-        /**
-         * 获取表达式的值
-         * @param element 目标对象
-         * @param property 表达式字符串
-         * @param data 数据上下文
-         */
-        private getValue(element: JQuery<HTMLElement>, property: string, data: any) {
-            let result = this.calcu(property, element, data);
-            let type = $.type(result);
-            if (type === 'string' || type === 'number' || type === 'boolean' || type === 'null' || type === 'undefined') {
-                return result;
-            } else {
-                return $.extend(this.toDefault(type), result);
-            }
-        }
-
-        /**
-         * 根据表达式字符串计算值
-         * @param property 表达式
-         * @param element 表达式所在的对象
-         * @param data 数据上下文
-         */
-        private calcu(property: string, element: JQuery<HTMLElement>, data: any) {
-            let nodes = property.match(/[@\w]+((?:\(.*?\))*|(?:\[.*?\])*)/g);
-            if (!nodes) {
-                return '';
-            }
-            let result = nodes.reduce((obj, current) => {
-                let arr = /^([@\w]+)([\(|\[].*)*/.exec(current);
-                let property = arr[1];
-                let tempData;
-                if (/^@.*/.test(property)) {
-                    tempData = this.getAliasData(element, property.substr(1));
-                }
-                else {
-                    tempData = obj[property];
-                }
-                let symbols = arr[2];
-                if (symbols) {
-                    let arr = symbols.match(/\[\d+\]|\(.*\)/g);
-                    return arr.reduce((obj2, current2) => {
-                        if (/\[\d+\]/.test(current2)) {
-                            let arrayIndex = parseInt(current2.replace(/\[(\d+)\]/, '$1'));
-                            return obj2[arrayIndex];
-                        }
-                        else if (/\(.*\)/.test(current2)) {
-                            let params = current2.replace(/\((.*)\)/, '$1').replace(/\s/, '').split(',');
-                            let computedParams = params.map(p => {
-                                if (/^[A-Za-z_\$\@].*$/.test(p)) {
-                                    return this.calcu(p, element, data);
-                                }
-                                else {
-                                    if (p === '') {
-                                        p = '""';
-                                    }
-                                    return (new Function('return ' + p))();
-                                }
-                            });
-                            let func: Function = obj2 || this.getMethod(property) || getMethod(property) || window[property];
-                            return func.apply(this, computedParams);
-                        }
-                    }, tempData);
-                }
-                else {
-                    return tempData;
-                }
-            }, data);
-            return result;
-        }
-
-        private toDefault(type: string) {
-            switch (type) {
-                case 'string':
-                    return '';
-                case 'number':
-                    return 0;
-                case 'boolean':
-                    return false;
-                case 'array':
-                    return [];
-                case 'object':
-                    return {};
-                case 'function':
-                    return function () { };
-                default:
-                    return null;
-            }
-        }
-
-        private setAlias(element: JQuery<HTMLElement>, property: string, data?: any) {
-            let targetData = data || element.data('context');
-            element.data('alias').self = {
-                data: targetData,
-                property: property
-            };
-            if (element.attr('p-as')) {
-                let alias = element.attr('p-as');
-                element.data('alias')[alias] = {
-                    data: targetData,
-                    property: property
-                }
-            }
-        }
-
-        private getAliasData(element: JQuery<HTMLElement>, alias: string) {
-            let data = element.data('alias');
-            return data[alias].data;
-        }
-
-        private getAliasProperty(element: JQuery<HTMLElement>, alias: string) {
-            let data = element.data('alias');
-            return data[alias].property;
-        }
-
-        private getMethod(name: string): Function {
+        public getMethod(name: string): Function {
             return this._methods[name];
         }
 
-        private dividePipe(expression: string) {
-            let array = expression.split('|');
-            let property = array[0].replace(/\s/g, '');
-            let method = array[1] ? array[1].replace(/\s/g, '') : null;
-            return {
-                property: property,
-                method: method
-            }
-        }
-
-        private convert(method: string, data: any) {
-            if (/^{.*}$/.test(method)) {
-                let pairs = method.replace(/{|}/g, '').split(',');
-                return pairs.reduce((pre, current) => {
-                    let pair = current.split(':');
-                    if (/^[a-zA-z$_]+/.test(pair[1])) {
-                        pre[pair[0]] = pair[1].split('.').reduce((predata, property) => {
-                            return predata[property];
-                        }, data);
-                    } else {
-                        pre[pair[0]] = new Function('return ' + pair[1])();
-                    }
-                    return pre;
-                }, {})
-            } else {
-                if (!hasSuffix(method, 'Converter')) {
-                    method += 'Converter';
-                }
-                return this._converters[method](data);
-            }
-        }
-
-        private filter(method: string, data: any[]) {
+        public filter(method: string, data: any[]) {
             if (!hasSuffix(method, 'Filter')) {
                 method += 'Filter';
             }
             return this._filters[method](data);
         }
+
+        
+
+        
+
+        
 
         public register(name: string, value: any) {
             if ($.isFunction(value)) {
